requirements.txt
streamlit==1.25.0
opencv-python==4.8.0
face_recognition==1.3.0
numpy==1.26.0
pandas==2.2.0
openpyxl==3.1.2
Pillow==10.0.0
